New Scene Tests
I attempted making a simple scene so that we can test things in it. I
imported a character controller as placeholder until Brent shows me how
to connect all the scripts properly.

In the test scene I made a simple "stapler" to test animation in Unity,
and it is a bit complicated. I'll work more on it, but the animation is
made but has no trigger for it.

Also, I want to see how the player might react to a moving object like
that and if we need to code it any differently.