CharacterAnimations
I successfully tied the character animations on running left and right
and being idle facing left and right.

It looks  pretty good but at times is a little slow to transition. I
will do some more looking into how to make the transitions more
responsive but if either of you guys know how to fix it then be my
guest.